Extract tell tests to test_tell.py
